trading-ai-companion/requirements.txt

### requirements.txt

ollama>=0.1.0
requests>=2.28.0
flask>=2.2.0
flask-socketio>=5.3.0
pyyaml>=6.0
pandas>=1.5.0
numpy>=1.24.0
websocket-client>=1.5.0
psutil>=5.9.0
pytest>=7.2.0
python-dotenv>=0.21.0
ta-lib>=0.4.24
matplotlib>=3.6.0
seaborn>=0.12.0
pyzmq>=25.0.0